+++ core/api/src/main/java/mb/pie/api/Pie.java
@@ -44,8 +44,9 @@ public interface Pie extends AutoCloseable {
      *
      * @param task Task to check. The {@link Task#key() key} of this task will be used to check.
      * @return True if task was executed at least once, false otherwise.
+     * @deprecated Use {@link Session#hasBeenExecuted(Task)}.
      */
-    default boolean hasBeenExecuted(Task<?> task) {
+    @Deprecated default boolean hasBeenExecuted(Task<?> task) {
         return hasBeenExecuted(task.key());
     }
 
@@ -54,121 +55,16 @@ public interface Pie extends AutoCloseable {
      *
      * @param key Key of task to check.
      * @return True if task was executed at least once, false otherwise.
+     * @deprecated Use {@link Session#hasBeenExecuted(Task)}.
      */
-    boolean hasBeenExecuted(TaskKey key);
-
-
-    /**
-     * Checks whether {@code task} is observed, either explicitly (by requiring it with a top-down build) or implicitly
-     * (when another observed task requires it).
-     *
-     * @param task Task to check. The {@link Task#key() key} of this task will be used to check.
-     * @return True if task is observed, false otherwise.
-     */
-    default boolean isObserved(Task<?> task) {
-        return isObserved(task.key());
-    }
-
-    /**
-     * Checks whether task with given {@code key} is observed, either explicitly (by requiring it with a top-down build)
-     * or implicitly (when another observed task requires it).
-     *
-     * @param key Key of task to check.
-     * @return True if task is observed, false otherwise.
-     */
-    boolean isObserved(TaskKey key);
-
-
-    /**
-     * Checks whether {@code task} is explicitly observed (by requiring it with a top-down build).
-     *
-     * @param task Task to check. The {@link Task#key() key} of this task will be used to check.
-     * @return True if task is explicitly observed, false otherwise.
-     */
-    default boolean isExplicitlyObserved(Task<?> task) {
-        return isExplicitlyObserved(task.key());
-    }
-
-    /**
-     * Checks whether task with given {@code key} is explicitly observed (by requiring it with a top-down build).
-     *
-     * @param key Key of task to check.
-     * @return True if task is explicitly observed, false otherwise.
-     */
-    boolean isExplicitlyObserved(TaskKey key);
-
-
-    /**
-     * Sets the observability of {@code task} to {@link Observability#ExplicitObserved explicitly observed} if it is
-     * {@link Observability#ImplicitObserved implicitly observed}. Does nothing if already {@link
-     * Observability#ExplicitObserved explicitly observed}. Throws if {@link Observability#Unobserved unobserved}. Use
-     * {@link Session#require} to explicitly observe an unobserved task.
-     *
-     * @param task Task to explicitly observe.
-     * @throws IllegalArgumentException when {@code} task is not observed.
-     */
-    default void setImplicitToExplicitlyObserved(Task<?> task) {
-        setImplicitToExplicitlyObserved(task.key());
-    }
-
-    /**
-     * Sets the observability of task with given {@code key} to {@link Observability#ExplicitObserved explicitly
-     * observed} if it is {@link Observability#ImplicitObserved implicitly observed}. Does nothing if already {@link
-     * Observability#ExplicitObserved explicitly observed}. Throws if {@link Observability#Unobserved unobserved}. Use
-     * {@link Session#require} to explicitly observe an unobserved task.
-     *
-     * @param key Key of task to explicitly observe.
-     * @throws IllegalArgumentException when {@code} task is not observed.
-     */
-    void setImplicitToExplicitlyObserved(TaskKey key);
-
-
-    /**
-     * Sets {@code function} as the callback for outputs of {@code task}. Whenever {@code task} is required, its
-     * up-to-date output will be passed as an argument to the {@code function}.
-     *
-     * @param task     Task to set the callback for. The {@link Task#key() key} of this task will be used to identify
-     *                 which callback function to call when a task is required.
-     * @param function Function to call with up-to-date output as argument when {@code task} is required. Consumed value
-     *                 by the function may be {@code null} when a task returns {@code null}.
-     */
-    <O extends Serializable> void setCallback(Task<O> task, Consumer<O> function);
-
-    /**
-     * Sets {@code function} as the callback for outputs of tasks with {@code key}. Whenever task with {@code key} is
-     * required, its up-to-date output will be passed as an argument to the {@code function}.
-     *
-     * @param key      Key of task to set callback for.
-     * @param function Function to call with up-to-date output as argument when task is required. Consumed value by the
-     *                 function may be {@code null} when a task returns {@code null}.
-     */
-    void setCallback(TaskKey key, Consumer<Serializable> function);
-
-    /**
-     * Removes the callback function for outputs of {@code task}.
-     *
-     * @param task Task to remove the callback for. The {@link Task#key()} of {@code task} will be used to identify
-     *             which callback function to remove.
-     */
-    void removeCallback(Task<?> task);
-
-    /**
-     * Removes the callback function for outputs of task with {@code key}.
-     *
-     * @param key Key of task to remove callback function for.
-     */
-    void removeCallback(TaskKey key);
-
-    /**
-     * Removes all callback functions.
-     */
-    void dropCallbacks();
+    @Deprecated boolean hasBeenExecuted(TaskKey key);
 
 
     /**
      * Removes all data from the store.
+     * @deprecated Use {@link Session#dropStore()}.
      */
-    void dropStore();
+    @Deprecated void dropStore();
 
 
     /**