+++ core/api/src/main/java/mb/pie/api/SerializableConsumer.java
@@ -1,0 +1,8 @@
+package mb.pie.api;
+
+import java.io.Serializable;
+import java.util.function.Consumer;
+import java.util.function.Supplier;
+
+@FunctionalInterface
+public interface SerializableConsumer<T extends Serializable> extends Consumer<T>, Serializable {}