+++ core/runtime/src/main/java/mb/pie/runtime/exec/BottomUpRunner.java
@@ -254,7 +254,7 @@ public class BottomUpRunner implements RequireTask {
             visited.put(key, storedData);
 
             // Invoke callback, if any.
-            final @Nullable Consumer<@Nullable Serializable> callback = callbacks.get(key);
+            final @Nullable Consumer<@Nullable Serializable> callback = callbacks.get(key, txn);
             if(callback != null) {
                 tracer.invokeCallbackStart(callback, key, output);
                 callback.accept(output);
@@ -330,7 +330,7 @@ public class BottomUpRunner implements RequireTask {
         visited.put(key, data);
 
         // Invoke callback, if any.
-        final @Nullable Consumer<@Nullable Serializable> callback = callbacks.get(key);
+        final @Nullable Consumer<@Nullable Serializable> callback = callbacks.get(key, txn);
         if(callback != null) {
             tracer.invokeCallbackStart(callback, key, data.output);
             callback.accept(data.output);