+++ core/runtime/src/main/java/mb/pie/runtime/exec/TaskExecutor.java
@@ -185,7 +185,7 @@ public class TaskExecutor {
         visited.put(key, data);
 
         // Invoke callback, if any.
-        final @Nullable Consumer<@Nullable Serializable> callback = callbacks.get(key);
+        final @Nullable Consumer<@Nullable Serializable> callback = callbacks.get(key, txn);
         if(callback != null) {
             tracer.invokeCallbackStart(callback, key, output);
             callback.accept(output);